Add hierarchical UI system
= Introduction =

This commit implements the beginnings of a UI widget system. The basic
idea is that the UI code should be abstracted somewhat from the rest of
crawl. This new code provides a small framework for two main components
of the UI: layout and rendering.

This has a number of advantages:

 - Resize support for free.
 - UI code is aware of the entire UI that needs to be drawn: this meansd
   no weird edge cases because the event handling loop in one function
   isn't (and can't) repaint something its caller needs repainted... etc.
 - No (fewer?) hacks: no peppering redraw() calls all over the place
   with wild abandon, no CK_REDRAW.

In addition, the same UI layout code is used for both SDL tiles and
console builds: the only difference is that tiles uses pixel coords,
and console uses terminal char coords. Rendering and layout for both
platforms is fully abstracted.

= Layout / Rendering =

get_preferred_size() retrieves a widget's minimum and "natural" size on
a given dimension (a "size request"). If retrieving the vertical size
request, a prospective width must be given, and the returned minimum and
natural size will be calculated according to that width: this allows a
text widget to adjust its required height depending on its width, by
rewrapping text to fit the width. Container widgets (those that contain
other widgets) will generally call this method on their children in
order to calculate their own size requirements.

When the UI hierarchy needs to be laid out, the horizontal minimum
and natural sizes are computed by calling get_preferred_size() on
the toplevel widget: this is a recursive process. The viewport
(screen/terminal) width is clamped to the minimum and natural widths,
and the result is used to compute the vertical size request, again using
get_preferred_size(). Likewise, the viewport height is clamped to the
result. The resulting viewport width and height are passed to
allocate_region().

Each container widget is responsible for laying out its children within
its allocated region, and then calling allocate_region() on each of them
in turn. Leaf widgets (text/image/etc.) usually pack buffers at this
point.

Rendering is relatively straight-forward: widgets render their contents
to their allocated regions, and then call render() on their children.

= API =

The main API is designed to be used with existing crawl code.
ui_pump_events() is a replacement for the ad-hoc while-getch loops seen
throughout the crawl codebase; as it has full knowledge of the UI
layout, it can render the full UI, including UI from code completely
unrelated to the call site.

This function is meant to be used in conjunction with event handlers
(implemented in a later commit).