Get ui-overhaul working on top of webtiles-changes
Ideally, these dependency changes would be fixup-ed into the relevent
commits, but that's not too likely as webtiles-changes is currently in
limbo and ui-overhaul is most likely going to stay based on the older
webtiles implementation.

Not too sure about the change to game_map.py, but it was necessary to
get webtiles to run for me.