Show ring swap menu when wearing from inventory
When putting on a ring from the inventory item description, the prompt
in the messages region was obscured by the UI shade. When putting on
rings in this way, just always show the menu; it's a simple fix and
more intuitive than hiding the popup menus.

It's still a little weird that (p)utting on the ring immediately
dismisses the description popup.