Fix webtiles menu scrolling with non-100% zoom (NormalPerson7)
NormalPerson7 noticed that menus got stuck when scrolling while using
certain zoom levels. This was caused by update_visible_indices() not
correctly accounting for the page zoom level, and thus always reporting
the same (incorrect) supposedly-visible region.

Since this function had a lot of handling for the previous
implementation, which supported partially loaded menus, I've replaced it
with a simpler version that is a reasonably straight-forward conversion
of the code in menu.cc.